requirements.txt
# Discord Todo 봇
discord.py>=2.3.0
python-dotenv>=1.0.0
supabase>=2.0.0
httpx>=0.27.0

openai>=1.0.0

# 개발
pytest>=8.0.0

# 레거시 RAG 실험: legacy/rag_stack — 별도 requirements-legacy.txt 권장
